+++ database/migrations/2020_06_18_182954_create_product_tags_table.php
@@ -16,8 +16,8 @@ class CreateProductTagsTable extends Migration
         Schema::dropIfExists('product_tags');
         Schema::create('product_tags', function (Blueprint $table) {
             $table->id();
-            $table->integer('product_id');
-            $table->integer('tag_id');
+            $table->bigInteger('product_id');
+            $table->bigInteger('tag_id');
         });
     }
